jQuery implementation
For the record, we are still using react to host/run the app, I did not change anything related to that, which means when you want to run the app, you still need to use npm start as before, the only thing I have changed is, I have moved the from logic ./src/index.js to ./public/index.html. Everything is the same except how the data is displayed to the DOM, to be specif, the code "$("#surveyElement").Survey({model: survey});" line number 481 instead of "ReactDOM.render(<Survey.Survey model={survey}/>, document.getElementById("surveyElement"));" which was in ./src/index.js line number 716.